% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genres.R
\name{genres}
\alias{genres}
\title{Genres}
\usage{
genres(file_name, song_column, song_name_in_quotations)
}
\arguments{
\item{file_name}{The name of the data.frame}

\item{song_column}{the column data of the song names of the given data.frame. Ex. dataset$track_name}

\item{song_name_in_quotations}{Song name. Ex. "Comedy"}
}
\description{
Identifies the cross-listed genres of a given song within the dataset.
}
\examples{
genres
}
\keyword{Spotify}
\keyword{genre}
\keyword{kaggle}
\keyword{search}
