% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Random_songs.R
\name{random_songs}
\alias{random_songs}
\title{Random_songs}
\usage{
random_songs(file_name, how_many, genre_in_quotations)
}
\arguments{
\item{file_name}{The name of the data.frame}

\item{how_many}{how many songs the user would like output Ex. 15}

\item{genre_in_quotations}{what genre from the given data.frame would the user like to output songs from Ex. "acoustic"}
}
\description{
Provides the user with a list of random songs given user genre preference
}
\examples{
random_songs
}
\keyword{Spotify}
\keyword{genre}
\keyword{kaggle}
\keyword{new}
\keyword{playlist}
\keyword{random}
