% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Search.R
\name{search}
\alias{search}
\title{Search}
\usage{
search(file_name, song_column, song_name_in_quotations)
}
\arguments{
\item{file_name}{The name of the data.frame}

\item{song_column}{the column data of the song names of the given data.frame. Ex. dataset$track_name}

\item{song_name_in_quotations}{Song name. Ex. "Comedy"}
}
\description{
Allows the user to search the data.frame for the qualities of a given song
}
\examples{
search
}
\keyword{Spotify}
\keyword{kaggle}
\keyword{playlist}
\keyword{qualities}
\keyword{search}
